requirements.txt
langgraph
langchain-core
langchain-groq
python-docx
PyMuPDF
python-dotenv
